fix: pr review comments
